Move newsletter subscription into newsletter routes
