+++ app/actions/blog/controller.test.ts
@@ -22,7 +22,7 @@ describe("Blog route", () => {
 
     expect(html).toContain("<title>Remix Blog</title>");
     expect(html).toContain('id="main-content"');
-    expect(html).toContain(`action="${routes.api.newsletter.href()}"`);
+    expect(html).toContain(`action="${routes.newsletter.subscribe.href()}"`);
 
     let articleImages = [
       ...html.matchAll(/<img\b(?:[^"'<>]|"[^"]*"|'[^']*')*>/g),
